Fix - cmd.p1 param (uint16_t)->(uint32_t) - param1~4 all can use as (uint8_t) at the same time

+++ libraries/AP_Mission/AP_Mission.h
@@ -400,7 +400,7 @@ public:
     struct Mission_Command {
         uint16_t index;             // this commands position in the command list
         uint16_t id;                // mavlink command id
-        uint16_t p1;                // general purpose parameter 1
+        uint32_t p1;                // general purpose parameter 1
         Content content;
 
         // for items which store in location, we offer a few more bits
